% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcECEst}
\alias{CalcECEst}
\title{Gross Removals from Enhancements}
\usage{
CalcECEst(FPTotal, RemEstAR)
}
\arguments{
\item{FPTotal}{Gross removals from Forest Plantations}

\item{RemEstAR}{Gross removals from Afforestation}
}
\value{
Gross Removals from Enhancements - tCO2e
}
\description{
This function uses equation (2) to calculate the Gross Removals
from Enhancements. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcRemTotalFP]{CalcRemTotalFP()}}

\code{\link[=CalcRemARTotal]{CalcRemARTotal()}}
}
