% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcEmEstTotalFP}
\alias{CalcEmEstTotalFP}
\title{Gross emissions Forest Plantations (Hard- and Softwood)}
\usage{
CalcEmEstTotalFP(EmEstFPHW, EmEstFPSW)
}
\arguments{
\item{EmEstFPHW}{Emissions from Hardwood Plantations}

\item{EmEstFPSW}{Emissions from Softwood Plantations}
}
\value{
Gross Emissions from Hard and Softwood Plantations
}
\description{
This function refers to the CalcEmForPlantHW and
CalcEmForPlantSW functions to calculate the gross emissions from forest
plantations. Gross Removals are presented in tCO2.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
\seealso{
\code{\link[=CalcEmForPlantHW]{CalcEmForPlantHW()}}

\code{\link[=CalcEmForPlantSW]{CalcEmForPlantSW()}}
}
