% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcEmForPlantHW}
\alias{CalcEmForPlantHW}
\title{Forest Plantations - Emissions from Hardwood Plantations}
\usage{
CalcEmForPlantHW(Volume, BioConvExp, RootToShoot)
}
\arguments{
\item{Volume}{Volume of hardwood plantation harvested (m^3)}

\item{BioConvExp}{Biomass conversion and expansion factor for hardwood
logging}

\item{RootToShoot}{Root-to-shoot ratio for tropical rainforests}
}
\value{
Emissions from Hardwood Plantations
}
\description{
This function refers to Equations (51-54) to calculate the gross
annual emissions from hardwood plantations. Note Equations 53 and 54 are
covered by the ConvBiomassToCo2e function. Emissions are presented in
tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
