% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deforestation.R
\name{CalcEmTotalDF}
\alias{CalcEmTotalDF}
\title{Emissions From Deforestation}
\usage{
CalcEmTotalDF(EmEstDeforUp, EmEstDeforLow)
}
\arguments{
\item{EmEstDeforUp}{Emission Estimate from Lowland Deforestation}

\item{EmEstDeforLow}{Emission Estimate from Upland Deforestation}
}
\value{
Net Emission from deforestation
}
\description{
This function references Equation (13) in order to calculate the
emissions from deforestation in upland and lowland areas. The resulting
emission is expressed in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.2}
}
\seealso{
\code{\link[=CalcEmDF]{CalcEmDF()}}
}
