% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deforestation.R
\name{CalcEstEmDefor}
\alias{CalcEstEmDefor}
\title{Emissions From Deforestation Lowland and Upland}
\usage{
CalcEstEmDefor(Area, EF)
}
\arguments{
\item{Area}{Area deforested in upland or lowland}

\item{EF}{Emissions Factor for deforestation in upland or lowland forest
tCO2e/ha}
}
\value{
Emission for deforestation in upland or lowland area
}
\description{
This function references Equations (11) and (12) in order to calculate the
emissions from deforestation in upland or lowland areas. The resulting
emission is expressed in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.2}
}
