% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Felling.R
\name{CalcEstEmFell}
\alias{CalcEstEmFell}
\title{Felling in Natural Forest}
\usage{
CalcEstEmFell(Volume, EF)
}
\arguments{
\item{Volume}{Volume of area logged (ha)}

\item{EF}{Total Emissions Factor (TEF)}
}
\value{
Emissions from Logging of Natural Forests - tCO2e
}
\description{
This function uses Equation (14) to calculate the gross emissions from logging
activities. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.3.1}
}
