% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burning.R
\name{CalcEstEmFire}
\alias{CalcEstEmFire}
\title{Emissions From Fire}
\usage{
CalcEstEmFire(
  Age,
  MAIBsw,
  RootToShootDryLandSmall,
  Area,
  local_CombustFactor = CombustFactor,
  local_GWP_CO2 = GWPCO2,
  local_EF_CO2 = EFCO2,
  local_GWP_CH4 = GWPCH4,
  local_EF_CH4 = EFCH4,
  local_GWP_N2O = GWPN2O,
  local_EF_N2O = EFN2O
)
}
\arguments{
\item{Age}{The age of the compartment that burnt in the year}

\item{MAIBsw}{The mean annual increment of above and below ground biomass in
softwood plantations}

\item{RootToShootDryLandSmall}{Root-to-shoot ratio for tropical moist
deciduous forest < 125 tB ha-1}

\item{Area}{Area burnt in softwood plantations in the year}
}
\value{
Emissions from Forest Degradation (Fire) - tCO2e
}
\description{
This function references Equations (20-27) to estimate
the forest degradation emissions due to fire for the reporting
year. The resulting value is expressed in tCO2e/yr.
}
\references{
\link{TBC - ERPD citation - Section 8.3.3.2}
}
