% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcEstEmRemsEnh}
\alias{CalcEstEmRemsEnh}
\title{Gross Removals from Enhancements}
\usage{
CalcEstEmRemsEnh(NetEmRemsFPln, EstRemARefor)
}
\arguments{
\item{NetEmRemsFPln}{Gross removals from Forest Plantations}

\item{EstRemARefor}{Gross removals from Afforestation}
}
\value{
Gross Removals from Enhancements - tCO2e
}
\description{
This function uses equation (2) to calculate the Gross Removals
from Enhancements. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcGrossRemFPln]{CalcGrossRemFPln()}}

\code{\link[=CalcGrossRemARefor]{CalcGrossRemARefor()}}
}
