% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcEstEmRemsFDeg}
\alias{CalcEstEmRemsFDeg}
\title{Forest Degradation Net Emissions}
\usage{
CalcEstEmRemsFDeg(EstEmFell, EstRemFell, EstEmFire, EstEmNFDeg)
}
\arguments{
\item{EstEmFell}{Gross emissions from logging}

\item{EstRemFell}{Gross removals from regrowth on logged areas}

\item{EstEmFire}{Gross emissions from burning}

\item{EstEmNFDeg}{Gross emissions from degradation}
}
\value{
Forest Degradation Net Emissions - tCO2e
}
\description{
This function uses equation (3) to calculate the net emissions
from Forest Degradation. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcEstEmFell]{CalcEstEmFell()}}

\code{\link[=CalcEstRemFell]{CalcEstRemFell()}}

\code{\link[=CalcEstEmFire]{CalcEstEmFire()}}

\code{\link[=CalcEstEmNFDeg]{CalcEstEmNFDeg()}}
}
