% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcEstEmRemsFDegNonProxy}
\alias{CalcEstEmRemsFDegNonProxy}
\title{Forest Degradation Non Proxy Net Emissions}
\usage{
CalcEstEmRemsFDegNonProxy(EstEmFire, EstEmNFDeg)
}
\arguments{
\item{EstEmFire}{Gross emissions from burning}

\item{EstEmNFDeg}{Gross emissions from degradation}
}
\value{
Forest Degradation Non Proxy Net Emissions - tCO2e
}
\description{
This function uses equation (3) to calculate the net emissions
from Forest Degradation. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcEstEmFire]{CalcEstEmFire()}}

\code{\link[=CalcEstEmNFDeg]{CalcEstEmNFDeg()}}
}
