% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Felling.R
\name{CalcEstRemFell}
\alias{CalcEstRemFell}
\title{Removals from regrowth on Felled Areas in Natural Forests}
\usage{
CalcEstRemFell(Area, MAIC, Age)
}
\arguments{
\item{Area}{Area of natural forest logged}

\item{MAIC}{Mean Annual Increment Carbon}
}
\value{
Removals from regrowth on Felled Areas in Natural Forests - tCO2e
}
\description{
This function uses Equation (16) to calculate the gross removals from regrowth
on felled areas in natural forest. Emissions are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.3.1}
}
