% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcFDEst}
\alias{CalcFDEst}
\title{Forest Degradation Net Emissions}
\usage{
CalcFDEst(EstEmFell, EstRemFell, EmFireTotal)
}
\arguments{
\item{EstEmFell}{Gross emissions from logging}

\item{EstRemFell}{Gross removals from regrowth on logged areas}

\item{EmFireTotal}{Gross emissions from burning}
}
\value{
Forest Degradation Net Emissions - tCO2e
}
\description{
This function uses equation (3) to calculate the net emissions
from Forest Degradation. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcEmFell]{CalcEmFell()}}

\code{\link[=CalcRemFell]{CalcRemFell()}}

\code{\link[=CalcEmFire]{CalcEmFire()}}
}
