% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcGrossEm}
\alias{CalcGrossEm}
\title{Gross Emissions}
\usage{
CalcGrossEm(GrossEmDefor, EstEmFell, EstEmFire, GrossEmFPln)
}
\arguments{
\item{GrossEmDefor}{Gross emissions from deforestation}

\item{EstEmFell}{Gross emissions from logging}

\item{EstEmFire}{Gross emissions from burning}

\item{GrossEmFPln}{Gross emissions from forest plantations}
}
\value{
Gross Emissions - tCO2e
}
\description{
This function uses equation (1) to calculate the gross emissions
from all sources. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcEstEmDefor]{CalcEstEmDefor()}}

\code{\link[=CalcEstEmFell]{CalcEstEmFell()}}

\code{\link[=CalcGrossEmFPln]{CalcGrossEmFPln()}}
}
