% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deforestation.R
\name{CalcGrossEmDefor}
\alias{CalcGrossEmDefor}
\title{Emissions From Deforestation}
\usage{
CalcGrossEmDefor(EstEmDeforUp, EstEmDeforLow)
}
\arguments{
\item{EstEmDeforUp}{Emission Estimate from Lowland Deforestation}

\item{EstEmDeforLow}{Emission Estimate from Upland Deforestation}
}
\value{
Net Emission from deforestation
}
\description{
This function references Equation (13) in order to calculate the
emissions from deforestation in upland and lowland areas. The resulting
emission is expressed in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.2}
}
\seealso{
\code{\link[=CalcEstEmDefor]{CalcEstEmDefor()}}
}
