% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcGrossEmFPln}
\alias{CalcGrossEmFPln}
\title{Gross emissions Forest Plantations (Hard- and Softwood)}
\usage{
CalcGrossEmFPln(EstEmFPlnHwd, EstEmFPlnSwd)
}
\arguments{
\item{EstEmFPlnHwd}{Emissions from Hardwood Plantations}

\item{EstEmFPlnSwd}{Emissions from Softwood Plantations}
}
\value{
Gross Emissions from Hard and Softwood Plantations
}
\description{
This function refers to theCalcEstEmFPlnHwd and
CalcEstEmFPlnSwd functions to calculate the gross emissions from forest
plantations. Gross Removals are presented in tCO2.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
\seealso{
\code{\link[=CalcEstEmFPlnHwd]{CalcEstEmFPlnHwd()}}

\code{\link[=CalcEstEmFPlnSwd]{CalcEstEmFPlnSwd()}}
}
