% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcGrossEmNoFDTotal}
\alias{CalcGrossEmNoFDTotal}
\title{Gross Emissions without degradation}
\usage{
CalcGrossEmNoFDTotal(EmEstDFTotal, EmEstFPTotal)
}
\arguments{
\item{EmEstDFTotal}{Gross emissions from deforestation}

\item{EmEstFPTotal}{Gross emissions from forest plantations}
}
\value{
Gross Emissions without degradation - tCO2e
}
\description{
This function uses equation (1) to calculate the gross emissions
from all sources apart from degradation. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcEmDF]{CalcEmDF()}}

\code{\link[=CalcEmEstTotalFP]{CalcEmEstTotalFP()}}
}
