% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcGrossEmTotal}
\alias{CalcGrossEmTotal}
\title{Gross Emissions}
\usage{
CalcGrossEmTotal(EstEmRemsDefor, EstEmFell, EmFireTotal, EmEstFPTotal)
}
\arguments{
\item{EstEmRemsDefor}{Gross emissions from deforestation}

\item{EstEmFell}{Gross emissions from logging}

\item{EmFireTotal}{Gross emissions from burning}

\item{EmEstFPTotal}{Gross emissions from forest plantations}
}
\value{
Gross Emissions - tCO2e
}
\description{
This function uses equation (1) to calculate the gross emissions
from all sources. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcEmDF]{CalcEmDF()}}

\code{\link[=CalcEmFell]{CalcEmFell()}}

\code{\link[=CalcEmEstTotalFP]{CalcEmEstTotalFP()}}
}
