% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Afforestation.R
\name{CalcGrossRemARefor}
\alias{CalcGrossRemARefor}
\title{Removals from Afforestation - Total of upland and lowland}
\usage{
CalcGrossRemARefor(AreaTotal, MAIV, BCEF, RootToShootRatio)
}
\arguments{
\item{AreaTotal}{Total Area of Afforestation over the period}

\item{MAIV}{Mean annual volume increment for afforestation/reforestation
m^3/hectare/year}

\item{BCEF}{Biomass Conversion and expansion factor for increments in humid
tropical natural forests}

\item{RootToShootRatio}{Root-to-shoot ratio for tropical forests}
}
\value{
Emission Removals from Afforestation - tCO2e
}
\description{
This function references Equation (34) to calculate the removals from
afforestation/reforestation for the year The resulting value is expressed in
tCO2e.Upland and lowland data was provided for the FRL but total
afforestation area, not aggregated for upland and lowland,  will be provided
for future reporting.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.1}
}
