% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcGrossRemFPln}
\alias{CalcGrossRemFPln}
\title{Gross removals from Forest Plantations (Hard- and Softwood)}
\usage{
CalcGrossRemFPln(EstRemFPlnHwd, EstRemFPlnSwd)
}
\arguments{
\item{EstRemFPlnHwd}{Removals from Softwood Plantations}

\item{EstRemFPlnSwd}{Removals from Softwood Plantations}
}
\value{
Gross Removals from Hard and Softwood Plantations
}
\description{
This function refers to the CalcEstRemFPlnHwd and
CalcEstRemFPlnSwd functions to calculate the gross removals from forest
plantations. Gross Removals are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
\seealso{
\code{\link[=CalcEstRemFPlnHwd]{CalcEstRemFPlnHwd()}}

\code{\link[=CalcEstRemFPlnSwd]{CalcEstRemFPlnSwd()}}
}
