% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcGrossRemNoFDTotal}
\alias{CalcGrossRemNoFDTotal}
\title{Gross Removals without degradation}
\usage{
CalcGrossRemNoFDTotal(RemEstAR)
}
\arguments{
\item{RemEstAR}{Gross removals from afforestation}
}
\value{
Gross Removals without degradation - tCO2e
}
\description{
This function uses equation (2) to calculate the gross removals
from all sources apart from degradation. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcRemARTotal]{CalcRemARTotal()}}
}
