% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcGrossRemTotal}
\alias{CalcGrossRemTotal}
\title{Gross Removals}
\usage{
CalcGrossRemTotal(RemEstFell, RemEstAR, RemEstFPTotal)
}
\arguments{
\item{RemEstFell}{Gross removals from degradation}

\item{RemEstAR}{Gross removals from afforestation}

\item{RemEstFPTotal}{Gross removals from forest plantations}
}
\value{
Gross Removals - tCO2e
}
\description{
This function uses equation (2) to calculate the gross removals
from all sources. Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcRemFell]{CalcRemFell()}}

\code{\link[=CalcRemARTotal]{CalcRemARTotal()}}

\code{\link[=CalcRemTotalFP]{CalcRemTotalFP()}}
}
