% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcNetEmRems}
\alias{CalcNetEmRems}
\title{Net Emissions}
\usage{
CalcNetEmRems(GrossEm, GrossRem)
}
\arguments{
\item{GrossEm}{Gross emissions}

\item{GrossRem}{Gross removals}
}
\value{
Net Emissions  - tCO2e
}
\description{
This function uses equation (3) to calculate net emissions.
Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcGrossEm]{CalcGrossEm()}}

\code{\link[=CalcGrossRem]{CalcGrossRem()}}
}
