% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Felling.R
\name{CalcNetEmRemsFell}
\alias{CalcNetEmRemsFell}
\title{Net emissions from Logging of Natural Forests (tCO2e)}
\usage{
CalcNetEmRemsFell(EstEmFell, EstRemFell)
}
\arguments{
\item{EstEmFell}{Emissions from logging}

\item{EstRemFell}{Removals from regrowth on felled areas}
}
\value{
Emissions from Logging of Natural Forests - tCO2e
}
\description{
This function used functions CalcEstEmFell and CalcEstRemFell to calculate net
emissions from logging. Emissions are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.3.1}
}
\seealso{
\code{\link[=CalcEstEmFell]{CalcEstEmFell()}}

\code{\link[=CalcEstRemFell]{CalcEstRemFell()}}
}
