% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcNetEmRemsFPln}
\alias{CalcNetEmRemsFPln}
\title{Net Emissions from Forest Plantations (Hard and Softwood)}
\usage{
CalcNetEmRemsFPln(GrossEmFPln, GrossRemFPln)
}
\arguments{
\item{GrossEmFPln}{Gross emissions from Soft and Hardwood Plantations}

\item{GrossRemFPln}{Gross removals from Soft and Hardwood Plantations}
}
\value{
Net emissions from Hard and Softwood Plantations
}
\description{
This function references functions CalcGrossEmFPln and
CalcGrossRemFPln to calculate net emissions from Hard and Softwood
Plantations. Values are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
\seealso{
\code{\link[=CalcGrossEmFPln]{CalcGrossEmFPln()}}

\code{\link[=CalcGrossRemFPln]{CalcGrossRemFPln()}}
}
