% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Degradation.R
\name{CalcNetEmRemsNFDeg}
\alias{CalcNetEmRemsNFDeg}
\title{Net emissions from Degradation of Natural Forests (tCO2e)}
\usage{
CalcNetEmRemsNFDeg(EstEmNFDeg)
}
\arguments{
\item{EstEmNFDeg}{Emissions from degradation}
}
\value{
Emissions from degradation of Natural Forests - tCO2e
}
\description{
This function used functions CalcEstEmNFDeg to calculate net
emissions from degradation Emissions are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation}
}
\seealso{
\code{\link[=CalcEstEmNFDeg]{CalcEstEmNFDeg()}}
}
