% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportTotals.R
\name{CalcNetEmTotal}
\alias{CalcNetEmTotal}
\title{Net Emissions}
\usage{
CalcNetEmTotal(GrossEmTotal, GrossRemTotal)
}
\arguments{
\item{GrossEmTotal}{Gross emissions}

\item{GrossRemTotal}{Gross removals}
}
\value{
Net Emissions  - tCO2e
}
\description{
This function uses equation (3) to calculate net emissions.
Emissions are presented in tco2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.1}
}
\seealso{
\code{\link[=CalcGrossEmTotal]{CalcGrossEmTotal()}}

\link{CalcGrossRemTotal}
}
