% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcRemForPlantHW}
\alias{CalcRemForPlantHW}
\title{Forest Plantations - Removals from Hardwood Plantations}
\usage{
CalcRemForPlantHW(
  AreaJustGrowsHW,
  AreaPlanted,
  AreaHarvested,
  MAIV,
  BioConvExpInc,
  RootToShootRatio
)
}
\arguments{
\item{AreaJustGrowsHW}{Initial area of forest at start of year}

\item{AreaPlanted}{Area planted during the year}

\item{AreaHarvested}{Area from area stocked which is harvested during the
year}

\item{MAIV}{Mean annual volume increment in hardwood plantations
m^3/hectare/year}

\item{BioConvExpInc}{Biomass conversion and expansion factor for increment
taken from IPCC, 2006, Vol 4}

\item{RootToShootRatio}{Root-to-shoot ratio for tropical forests}
}
\value{
Removals from Hardwood Plantations
}
\description{
This function references Equations (56,57) & (60-63) and to calculate
the removals from Hardwood Plantations. Note Equations (62) and (63) are
covered by the ConvBiomassToCo2e function. Removals are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
