% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcRemForPlantSW}
\alias{CalcRemForPlantSW}
\title{Forest Plantations - Removals from Softwood Plantations}
\usage{
CalcRemForPlantSW(MAIBsw, AreaJustGrowsSW, AreaPlanted, AreaHarvested)
}
\arguments{
\item{MAIBsw}{Mean annual biomass increment in Softwood Plantations}

\item{AreaJustGrowsSW, }{Area stocked - area harvested in that year}

\item{AreaPlanted, }{Area planted during the year}

\item{AreaHarvested}{Area from area stocked which is harvested during the
year}
}
\value{
Removals from Softwood Plantations
}
\description{
This function references Equations (41), (42), (46), (47-49) to
calculate the removals from Softwood Plantations. Note Equations (48) and
(49) are covered by the ConvBiomassToCo2e function. Removals are presented
in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
