% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcRemTotalFP}
\alias{CalcRemTotalFP}
\title{Gross removals from Forest Plantations (Hard- and Softwood)}
\usage{
CalcRemTotalFP(RemEstFPHW, RemEstFPSW)
}
\arguments{
\item{RemEstFPHW}{Removals from Softwood Plantations}

\item{RemEstFPSW}{Removals from Softwood Plantations}
}
\value{
Gross Removals from Hard and Softwood Plantations
}
\description{
This function refers to the CalcRemForPlantHW and
CalcRemForPlantSW functions to calculate the gross removals from forest
plantations. Gross Removals are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
\seealso{
\code{\link[=CalcRemForPlantHW]{CalcRemForPlantHW()}}

\code{\link[=CalcRemForPlantSW]{CalcRemForPlantSW()}}
}
