% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestPlantations.R
\name{CalcTotalFP}
\alias{CalcTotalFP}
\title{Net Emissions from Forest Plantations (Hard and Softwood)}
\usage{
CalcTotalFP(EmEstFPTotal, RemEstFPTotal)
}
\arguments{
\item{EmEstFPTotal}{Gross emissions from Soft and Hardwood Plantations}

\item{RemEstFPTotal}{Gross removals from Soft and Hardwood Plantations}
}
\value{
Net emissions from Hard and Softwood Plantations
}
\description{
This function references functions CalcEmEstTotalFP and
CalcRemTotalFP to calculate net emissions from Hard and Softwood
Plantations. Values are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.4.2}
}
\seealso{
\code{\link[=CalcEmEstTotalFP]{CalcEmEstTotalFP()}}

\code{\link[=CalcRemTotalFP]{CalcRemTotalFP()}}
}
