% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarbonConversion.R
\name{ConvBiomassToCarbon}
\alias{ConvBiomassToCarbon}
\title{Biomass to Carbon Conversion}
\usage{
ConvBiomassToCarbon(Biomass)
}
\arguments{
\item{Biomass}{Tonnes of biomass dry matter (tB)}
}
\value{
Converting Biomass to Carbon
}
\description{
This function uses the reference parameters in Table 2 to convert tonnes of
dry biomass to Carbon. All values are presented in tonnes.
}
\references{
\link{TBC - ERPD citation - Section 8: Table 2}
}
