% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarbonConversion.R
\name{ConvBiomassToCO2e}
\alias{ConvBiomassToCO2e}
\title{Convert Biomass to CO2e}
\usage{
ConvBiomassToCO2e(Biomass)
}
\arguments{
\item{Biomass}{Tonnes of biomass dry matter (tB)}
}
\value{
Converting Biomass to CO2e
}
\description{
This function uses the defined functions ConvBiomassToCarbon and
ConvCarbonToCO2e to convert tonnes of dry biomass directly to Carbon Dioxide
equivalent values. All values are presented in tonnes.
}
\references{
\link{TBC - ERPD citation - Section 8}
}
\seealso{
\code{\link[=ConvBiomassToCarbon]{ConvBiomassToCarbon()}}

\code{\link[=ConvCarbonToCO2e]{ConvCarbonToCO2e()}}
}
