% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarbonConversion.R
\name{ConvCarbonToCO2e}
\alias{ConvCarbonToCO2e}
\title{Carbon to CO2e Conversion}
\usage{
ConvCarbonToCO2e(Carbon)
}
\arguments{
\item{Carbon}{Tonnes of Carbon (tC)}
}
\value{
Converting Carbon to CO2e
}
\description{
This function uses the reference parameters in Table 2 to convert Carbon to
Carbon Dioxide equivalent values. All values are presented in tonnes.
}
\references{
\link{TBC - ERPD citation - Section 8: Table 2}
}
