% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FijiNFMSCalculations-package.R
\docType{package}
\name{FijiNFMSCalculations-package}
\alias{FijiNFMSCalculations}
\alias{FijiNFMSCalculations-package}
\title{FijiNFMSCalculations: Fiji NFMS Calculations}
\description{
This package has tested functions for Fiji's NFMS Calculations
}
\author{
\strong{Maintainer}: Michael Green \email{github@ftt6283.org}

}
\keyword{internal}
