% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Felling.R
\name{CalcEmTotalFell}
\alias{CalcEmTotalFell}
\title{Net emissions from Logging of Natural Forests (tCO2e)}
\usage{
CalcEmTotalFell(EmEstFell, RemEstFell)
}
\arguments{
\item{EmEstFell}{Emissions from logging}

\item{RemEstFell}{Removals from regrowth on felled areas}
}
\value{
Emissions from Logging of Natural Forests - tCO2e
}
\description{
This function used functions CalcEMFell and CalcREMFell to calculate net
emissions from logging. Emissions are presented in tCO2e.
}
\references{
\link{TBC - ERPD citation - Section 8.3.3.1}
}
\seealso{
\code{\link[=CalcEmFell]{CalcEmFell()}}

\code{\link[=CalcRemFell]{CalcRemFell()}}
}
